Clarify help for `arg_many()` and `arg_many1()`
When displaying help the usage for both `arg_many()` and `arg_many1()` was
rendered as `ARG...`.  However, `...` is commonly used to mean "one or more":

- [POSIX.1-2024 12.1.9](https://pubs.opengroup.org/onlinepubs/9799919799/basedefs/V1_chap12.html#tag_12_01)
- [docopt](http://docopt.org/)

Hence for rendering `arg_many()`'s usage of "zero or more" something like
`[ARG...]` or `[ARG]...` should be used.  This would be in line with various
manual pages of `cat`, for example:

- [GNU cat](https://www.gnu.org/software/coreutils/manual/html_node/cat-invocation.html)
- [FreeBSD cat](https://man.freebsd.org/cgi/man.cgi?query=cat)
- [illumos cat](https://illumos.org/man/1/cat)
- [POSIX cat](https://pubs.opengroup.org/onlinepubs/9799919799/utilities/cat.html)

The usage of `arg_many()` is now rendered as `[ARG...]` while `arg_many1()`
remains as `ARG...`.